Replaced Enemy Class
Renamed/replaced the old Enemy class. It is now called "EnemyMovement".
